Fix bug report template
This commit adjusts the bug report template to ensure it works as
intended. It also documents an attempt to prevent the bug report from
opening in a modal. While using a direct link like
`issues/new?template=bug-report.md` can open the template in
full-screen mode, this approach results in the template appearing
twice in the issue creation list.

Unfortunately, there is no available option in GitHub to hide the bug
report template from the list while still allowing it to be accessible
via a direct link. This limitation has been noted, and the current
configuration is the best achievable under these constraints.

Close #8690